requirements.txt
Celery
django>=2.0
django-graphql-jwt
django-cors-headers
djangorestframework
docx
python-docx
graphene-django
psycopg2-binary
Scrapy
scrapyd
python-scrapyd-api
selenium
Spacy
pillow
Pypandoc
pyvirtualdisplay
